requirements.txt
pandas
numpy
IPython
scikit-learn
unidecode
pyenchant
twython==3.5.0
pymongo[tls,srv]==3.7.2

